requirements.txt
aiohttp==3.6.2
apipkg==1.5
APScheduler==3.6.3
asttokens==2.0.3
async-timeout==3.0.1
attrs==19.3.0
base58==2.0.1
bitarray==1.2.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cli==0.1.0
click==7.1.2
cryptography==3.2.1
cytoolz==0.10.1
-e git+https://bitbucket.org/nexid-edge-assets/db.git@919b19aacd9747ed41bfae009efdf314729e5002#egg=db
decorator==4.4.2
eth-abi==2.1.1
eth-account==0.5.2
eth-brownie==1.9.4
eth-event==1.1.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.9.0
execnet==1.7.1
gunicorn==20.0.4
hexbytes==0.2.1
hypothesis==5.16.2
idna==2.9
inflection==0.5.0
ipfshttpclient==0.4.13.2
jsonschema==3.2.0
logger==0.2.0
lru-dict==1.1.6
more-itertools==8.4.0
multiaddr==0.0.9
multidict==4.7.6
mysqlclient==2.0.1
mythx-models==1.9.1
netaddr==0.7.20
numpy==1.19.4
packaging==20.4
parsimonious==0.8.1
pluggy==0.13.1
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
py==1.8.2
py-solc-ast==1.2.4
py-solc-x==0.10.0
pycparser==2.20
pycryptodome==3.9.7
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
pytest-forked==1.1.3
pytest-xdist==1.32.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-telegram-bot==13.0
pythx==1.6.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
rlp==1.2.0
semantic-version==2.8.5
six==1.15.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.20
termcolor==1.1.0
terminaltables==3.1.0
toolz==0.10.0
tornado==6.1
tqdm==4.46.1
tzlocal==2.1
urllib3==1.25.9
utils==0.1.1
uvloop==0.14.0
varint==1.0.2
vyper==0.1.0b17
wcwidth==0.2.4
web3==5.11.1
websockets==8.1
yarl==1.4.2
